% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist.cohort.R
\name{exist.cohort}
\alias{exist.cohort}
\title{Function to extract if a cohort exists}
\usage{
exist.cohort(population, cohort)
}
\arguments{
\item{population}{Population list}

\item{cohort}{Cohort to check if it is contained in the population list}
}
\value{
TRUE/FALSE
}
\description{
Function to extract if a cohort exists
}
\examples{
data(ex_pop)
exist.cohort(ex_pop, cohort = "StrangeName_42")
exist.cohort(ex_pop, cohort = "Cohort_1_M")
}
