% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.qtl.R
\name{get.qtl}
\alias{get.qtl}
\title{QTL extraction}
\usage{
get.qtl(population)
}
\arguments{
\item{population}{Population list}
}
\value{
Vector of SNP positions
}
\description{
Function to the position of QTLs (for snp/chr use get.qtl.effects()
}
\examples{
data(ex_pop)
positions <- get.qtl(ex_pop)
}
