% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlist.R
\name{vlist}
\alias{vlist}
\title{Generation of a sublist}
\usage{
vlist(list, skip = NULL, first = NULL, select = NULL)
}
\arguments{
\item{list}{list you want to print details of}

\item{skip}{Skip first that many list-elements}

\item{first}{Only display first that many list-elements}

\item{select}{Display only selected list-elements}
}
\value{
Selected elements of a list
}
\description{
Internal function to write a couple of list entries in a new list
}
\examples{
data(ex_pop)
vlist(ex_pop$breeding[[1]], select=3:10)
}
